Fixed some refactoring issues
