+++ assets/src/types/notice.ts
@@ -3,7 +3,7 @@
  */
 import type { ReactNode } from 'react';
 
-export type NoticeKind = 'success' | 'error' | 'warning' | 'info';
+type NoticeKind = 'success' | 'error' | 'warning' | 'info';
 
 export interface NoticeState {
 	type: NoticeKind;
